Wrap up

+++ app/src/main/java/ibrahim/example/stocklogger/fragments/SellStockFragment.java
@@ -10,6 +10,7 @@ import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
 import androidx.preference.PreferenceManager;
 
+import android.util.Log;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -20,6 +21,8 @@ import android.widget.DatePicker;
 import android.widget.EditText;
 import android.widget.TextView;
 
+import com.daimajia.androidanimations.library.Techniques;
+import com.daimajia.androidanimations.library.YoYo;
 import com.google.android.material.snackbar.Snackbar;
 
 import java.text.DateFormat;
@@ -27,6 +30,7 @@ import java.util.Calendar;
 
 import ibrahim.example.stocklogger.R;
 import ibrahim.example.stocklogger.databases.StockDatabase;
+import ibrahim.example.stocklogger.pojos.ActiveStock;
 import ibrahim.example.stocklogger.pojos.SoldStock;
 import ibrahim.example.stocklogger.pojos.Stock;
 
@@ -47,6 +51,8 @@ public class SellStockFragment extends Fragment  {
     private String mParam2;
 
     public static TextView calendarTextDate;
+
+    EditText soldPriceEditText;
 
     public SellStockFragment() {
         // Required empty public constructor
@@ -86,7 +92,7 @@ public class SellStockFragment extends Fragment  {
         View view = inflater.inflate(R.layout.fragment_sell_stock, container, false);
 
         TextView soldNameTextView = view.findViewById(R.id.soldNameTextView);
-        EditText soldPriceEditText = view.findViewById(R.id.soldPriceEditText);
+        soldPriceEditText = view.findViewById(R.id.soldPriceEditText);
         EditText soldQuantityEditText = view.findViewById(R.id.soldQuantityEditText);
         calendarTextDate = view.findViewById(R.id.calendarTextDate);
         Button soldButton = view.findViewById(R.id.soldButton);
@@ -96,6 +102,7 @@ public class SellStockFragment extends Fragment  {
         soldPriceEditText.setOnEditorActionListener(new TextView.OnEditorActionListener() {
             @Override
             public boolean onEditorAction(TextView textView, int i, KeyEvent keyEvent) {
+                Log.d("TEST_RESPONSE", "Action ID = " + i + "KeyEvent = " + keyEvent);
                 return false;
             }
         });
@@ -127,41 +134,82 @@ public class SellStockFragment extends Fragment  {
 
                     if(soldPriceEditText.getText().toString().equals("")){
                         error = true;
+                        YoYo.with(Techniques.Shake)
+                                .duration(700)
+                                .playOn(soldPriceEditText);
                         Snackbar.make(view, "Sold price can't be blank", Snackbar.LENGTH_LONG)
                                 .setAction("Action", null).show();
                     }
 
-                    if(!error){
+                    if(soldQuantityEditText.getText().toString().equals("")){
+                        error = true;
+                        YoYo.with(Techniques.Shake)
+                                .duration(700)
+                                .playOn(soldQuantityEditText);
+                        Snackbar.make(view, "Sold quantity can't be blank", Snackbar.LENGTH_LONG)
+                                .setAction("Action", null).show();
+                    }
 
+                    if(!error){
                         double soldPrice = Double.parseDouble(soldPriceEditText.getText().toString());
+                        int soldQuantity = Integer.parseInt(soldQuantityEditText.getText().toString());
                         double worth = stock.getWorth();
                         double rating = 1;
 
-                        if(stock.isUSD()){
-                            SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(getContext());
-                            rating = Double.parseDouble(sp.getString("isUSD","1.26"));
+
+                        // If sold part of a stock
+                        if(soldQuantity<stock.getQuantity()){
+                            // get trading fee
+                            double tradingFee = 6.95;
+                            SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(getContext());
+                            tradingFee = Double.parseDouble(sharedPreferences.getString("tradingFee", "6.95"));
+
+                            // add a new record to active stock
+                            StockDatabase db = new StockDatabase(getContext());
+                            ActiveStock activeStock = new ActiveStock(
+                                    stock.getSymbol(),
+                                    stock.getCompanyName(),
+                                    soldPrice,
+                                    soldQuantity * -1,
+                                    calendarTextDate.getText().toString()
+                            );
+                            int activeStockId = db.addActiveStock(activeStock);
+
+                            // Update stock worth and quantity
+                            double soldWorth = (activeStock.getPrice() * activeStock.getQuantity()+ tradingFee) / activeStock.getQuantity();
+                            stock.setWorth(worth - soldWorth);
+                            stock.setQuantity(stock.getQuantity() - soldQuantity);
+                            db.updateStock(stock);
+
+                            // add a new relationship to stock_active table
+                            db.addStockActive(stock.getId(), activeStockId);
+                            db.close();
+                        } else {
+                            // If sold all holding stocks of a company
+                            if (stock.isUSD()) {
+                                SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(getContext());
+                                rating = Double.parseDouble(sp.getString("isUSD", "1.26"));
+                            }
+
+
+                            double earned = (soldPrice - worth) * soldQuantity * rating;
+
+                            StockDatabase db = new StockDatabase(getContext());
+
+                            db.addSoldStock(
+                                    new SoldStock(
+                                            stock.getSymbol(),
+                                            stock.getCompanyName(),
+                                            soldPrice,
+                                            earned,
+                                            calendarTextDate.getText().toString()
+                                    )
+                            );
+
+                            db.deleteStock(stock.getId());
+
+                            db.close();
                         }
-
-                        int quantity = Integer.parseInt(soldQuantityEditText.getText().toString());
-
-                        double earned = (soldPrice - worth) * quantity * rating;
-
-                        StockDatabase db = new StockDatabase(getContext());
-
-                        db.addSoldStock(
-                                new SoldStock(
-                                        stock.getSymbol(),
-                                        stock.getCompanyName(),
-                                        soldPrice,
-                                        earned,
-                                        calendarTextDate.getText().toString()
-                                )
-                        );
-
-                        db.deleteStock(stock.getId());
-
-                        db.close();
-                        soldPriceEditText.onEditorAction(EditorInfo.IME_ACTION_DONE);
                         Navigation.findNavController(view).popBackStack();
                     }
                 }
@@ -174,5 +222,10 @@ public class SellStockFragment extends Fragment  {
         return view;
     }
 
-
+    @Override
+    public void onDestroyView() {
+        Log.d("FRAGMENT","Destroyed");
+        soldPriceEditText.onEditorAction(EditorInfo.IME_ACTION_DONE);
+        super.onDestroyView();
+    }
 }
